moves api call to proper file structure
